mcp_project/mcp-server-project/pyproject.toml
[project]
name = "mcp-server-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=1.5.0",
    "anthropic>=0.51.0",
    "beautifulsoup4>=4.13.4",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-core>=0.3.59",
    "langchain-mcp-adapters>=0.0.11",
    "langchain-openai>=0.3.16",
    "langgraph>=0.4.3",
    "lxml>=5.4.0",
    "mcp[cli]>=1.8.0",
    "python-dotenv>=1.1.0",
    "serpapi>=0.1.5",
]
